do not entity encode the conversation title
This was initially done throughout a few different sections of the app to prevent xss. Since the title of the conversation is always defined by the user, I am not worried about this, here.

I could re-implement the UI side to support the entity encoding, but since the title is used throughout many different places (app bar, notifications, document title, settings, etc), it is probably easier to just remove this. Since safety isn't really an issue with it being defined by the user, this is fine.

closes #41