update the emoji set for the picker.
The emojione set is missing a ton of emojis, which just leaves blank spaces. A lot of the available sets were missing emojis. Twitter seems to be the closest to emojione and it seems to be the most complete. This will have to do, for now.

That means that the app could use three different types of emojis...
1. the picker is twitter
2. after selecting the emoji, it uses the system default when sitting in the sendbar
3. when the messages are rendered, it uses emojione